Solve day 18 part two
We hope, and verify, that picking up all keys in each
quarter is possible along the shortest route where we
ignore all doors. This is OK if it is possible to pick up
the relevant key from the same or another quadrant
earlier than needing any keys that are still behind this
door. Turns out this is valid for the puzzle input.