Make day 15 more general and refactor it
Explore the maze in part one in order of distance from
start rather than distance from where the droid is.
This is less efficient in terms of droid movement, but
ensures that the first path to oxygen is indeed the shortest
possible, and also ensures that it is found even if the maze
would be open ended.

Additionally refactor the code to allow more reuse between the
first and second parts.